requirements.txt
streamlit
google-generativeai>=0.8.5
pandas